refactor : DTO 위치를 query 디렉토리로 이동
