+++ src/main/java/com/example/bak/company/presentation/CompanyController.java
@@ -1,9 +1,9 @@
 package com.example.bak.company.presentation;
 
-import com.example.bak.company.application.CompanyService;
 import com.example.bak.company.application.command.CompanyCommandService;
-import com.example.bak.company.application.dto.CompanyResult;
-import com.example.bak.company.application.dto.CompanyResult.CompanyId;
+import com.example.bak.company.application.query.CompanyQueryService;
+import com.example.bak.company.application.query.dto.CompanyResult;
+import com.example.bak.company.application.query.dto.CompanyResult.CompanyId;
 import com.example.bak.company.presentation.dto.CompanyRequest;
 import com.example.bak.global.common.response.ApiResponse;
 import com.example.bak.global.common.response.ApiResponseFactory;
@@ -26,11 +26,11 @@ import org.springframework.web.bind.annotation.RestController;
 public class CompanyController {
 
     private final CompanyCommandService companyCommandService;
-    private final CompanyService companyService;
+    private final CompanyQueryService companyQueryService;
 
     @GetMapping()
     public ResponseEntity<ApiResponse> getCompanies() {
-        List<CompanyResult.Detail> companyResults = companyService.getCompanies();
+        List<CompanyResult.Detail> companyResults = companyQueryService.getCompanies();
         ApiResponse response = ApiResponseFactory.success(
                 "회사 목록을 성공적으로 조회했습니다.", companyResults);
 
@@ -41,7 +41,7 @@ public class CompanyController {
     public ResponseEntity<ApiResponse> getCompany(
             @PathVariable Long companyId
     ) {
-        CompanyResult.Detail companyResult = companyService.getCompany(companyId);
+        CompanyResult.Detail companyResult = companyQueryService.getCompany(companyId);
         ApiResponse response = ApiResponseFactory.success("회사 정보를 성공적으로 조회했습니다.", companyResult);
 
         return ResponseEntity.ok(response);
